python/revng/internal/pyproject.toml
# This file is distributed under the MIT License. See LICENSE.md for details.

[project]
name = "revng_internal"
description = "Internal revng modules"
dynamic = ["version"]

dependencies = [
  # Requirements for revng_merge_dynamic
  "pyelftools",
  # revng.api
  "cffi",
  # revng.daemon
  "ariadne",
  "uvicorn",
  "python-multipart",
  "starlette",
  "xdg",
  # CLI: revng graphql
  "aiohttp",
  "gql",
  "psutil",
  # CLI: revng model fetch-debuginfo
  "requests",
  "requests-toolbelt",
  "pefile",
  # CLI: revng model import idb
  "python-idb",
  # CLI: revng model compare
  "networkx",
  "grandiso",
  # General dependencies
  "revng",
  "PyYAML",
  "Jinja2",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
revng = [
  "internal/py.typed",
  "internal/daemon.yml",
  "internal/pypeline.yml",
  "internal/daemon/schema.graphql",
  "internal/daemon2/config_schema.json",
  "internal/cli/_commands/generate_migrations/metaschema.yml",
  "internal/cli/_commands/generate_migrations/migration.py.tpl",
  "internal/cli/_commands/llm_rename/prompt.tpl",
]

[project.entry-points.console_scripts]
revng = "revng.internal.cli.revng:main"
revng2 = "revng.internal.cli.revng2:run"

[tool.setuptools.dynamic]
version = {attr = "revng.internal.__version__"}
